+++ resources/lib/generators/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+# pylint: disable=missing-module-docstring
+from .epg_generator import EPGGenerator
+from .playlist_generator import PlaylistGenerator