+++ resources/lib/iptvmanager.py
@@ -1,19 +1,20 @@
 # -*- coding: utf-8 -*-
 """IPTV Manager Integration module"""
-from datetime import datetime
 import json
 import socket
+
+from lib.providers import ProviderInterface
 
 class IPTVManager:
     """IPTV Manager interface"""
 
-    def __init__(self, port, **kwargs):
+    def __init__(self, port: int, provider: ProviderInterface):
         """Initialize IPTV Manager object"""
         self.port = port
-        self.fetch_channels = kwargs.get('channels_loader')
-        self.fetch_programs = kwargs.get('programs_loader')
+        self.provider = provider
 
-    def via_socket(func): # pylint: disable=no-self-argument
+    # pylint: disable=no-self-argument
+    def via_socket(func):
         """Send the output of the wrapped function to socket"""
 
         def send(self):
@@ -28,59 +29,11 @@ class IPTVManager:
         return send
 
     @via_socket
-    def send_channels(self):
+    def send_channels(self) -> dict:
         """Return JSON-STREAMS formatted python datastructure to IPTV Manager"""
-        channels = self.fetch_channels()
-        streams = []
-
-        for channel in channels:
-            streams.append({
-                'id': channel['id'],
-                'name': channel['name'],
-                'preset': channel['zappingNumber'],
-                'logo': channel['logos']['square'],
-                'stream': 'plugin://plugin.video.orange.fr/channel/{id}'.format(id=channel['id'])
-            })
-
-        return { 'version': 1, 'streams': streams }
+        return { 'version': 1, 'streams': self.provider.get_streams() }
 
     @via_socket
-    def send_epg(self):
+    def send_epg(self) -> dict:
         """Return JSON-EPG formatted python data structure to IPTV Manager"""
-        programs = self.fetch_programs()
-        epg = {}
-
-        for program in programs:
-            if not program['channelId'] in epg:
-                epg[program['channelId']] = []
-
-            start = datetime.fromtimestamp(program['diffusionDate']).astimezone().replace(microsecond=0)
-            stop = datetime.fromtimestamp(program['diffusionDate'] + program['duration']).astimezone()
-
-            if program['programType'] != 'EPISODE':
-                title = program['title']
-                subtitle = None
-                episode = None
-            else:
-                title = program['season']['serie']['title']
-                subtitle = program['title']
-                episode = 'S{s}E{e}'.format(s=program['season']['number'], e=program['episodeNumber'])
-
-            image = None
-            if isinstance(program['covers'], list):
-                for cover in program['covers']:
-                    if cover['format'] == 'RATIO_16_9':
-                        image = program['covers'][0]['url']
-
-            epg[program['channelId']].append({
-                'start': start.isoformat(),
-                'stop': stop.isoformat(),
-                'title': title,
-                'subtitle': subtitle,
-                'episode': episode,
-                'description': program['synopsis'],
-                'genre': program['genre'] if program['genreDetailed'] is None else program['genreDetailed'],
-                'image': image
-            })
-
-        return { 'version': 1, 'epg': epg }
+        return { 'version': 1, 'epg': self.provider.get_epg(days=6) }