+++ resources/lib/providers/fr/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 # pylint: disable=missing-module-docstring
+from .orange import OrangeFranceProvider