+++ resources/lib/types.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""TYPES"""
+from enum import Enum
+
+class DRM(Enum):
+    """DRM"""
+    WIDEVINE = 'com.widevine.alpha'